LPS-32434 Test Formatting
