+++ portal-web/test/functional/com/liferay/portalweb/portlet/tagsnavigation/portlet/addportletstn/AddPortletsTNTests.html
@@ -9,9 +9,9 @@
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
 <tr><td><b>Test Suite</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/tagsnavigation/portlet/addportlettn/AddPageTNTest.html">AddPageTNTest</a></td></tr>
-<tr><td><a href="AddPortletTN1Test.html">AddPortletTN1Test</a></td></tr>
-<tr><td><a href="AddPortletTN2Test.html">AddPortletTN2Test</a></td></tr>
-<tr><td><a href="AddPortletTN3Test.html">AddPortletTN3Test</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/tagsnavigation/portlet/addportlettn/AddPortletTN1Test.html">AddPortletTN1Test</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/tagsnavigation/portlet/addportlettn/AddPortletTN2Test.html">AddPortletTN2Test</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/tagsnavigation/portlet/addportlettn/AddPortletTN3Test.html">AddPortletTN3Test</a></td></tr>
 <tr><td><a href="ViewPortletsTNTest.html">ViewPortletsTNTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
