+++ portal-web/test/functional/com/liferay/portalweb/portlet/tagsnavigation/portlet/addportlettn/AddPortletTNTests.html
@@ -10,6 +10,7 @@
 <tr><td><b>Test Suite</b></td></tr>
 <tr><td><a href="AddPageTNTest.html">AddPageTNTest</a></td></tr>
 <tr><td><a href="AddPortletTNTest.html">AddPortletTNTest</a></td></tr>
+<tr><td><a href="ViewPortletTNTest.html">ViewPortletTNTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
 </body>
