+++ portal-web/test/functional/com/liferay/portalweb/portlet/tagsnavigation/portlet/addportlettn/AddPortletTNTests.java
@@ -28,6 +28,7 @@ public class AddPortletTNTests extends BaseTestSuite {
 		TestSuite testSuite = new TestSuite();
 		testSuite.addTestSuite(AddPageTNTest.class);
 		testSuite.addTestSuite(AddPortletTNTest.class);
+		testSuite.addTestSuite(ViewPortletTNTest.class);
 		testSuite.addTestSuite(TearDownPageTest.class);
 
 		return testSuite;
