+++ portal-web/test/functional/com/liferay/portalweb/portlet/tagsnavigation/portlet/addportlettn/ViewPortletTNTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ViewPortletsTNTest</title>
+<title>ViewPortletTNTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">ViewPortletsTNTest</td></tr>
+<tr><td rowspan="1" colspan="3">ViewPortletTNTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
@@ -20,32 +20,37 @@
 </tr>
 <tr>
 	<td>assertVisible</td>
-	<td>//div[1]/section</td>
+	<td>//section</td>
 	<td>Portlet</td>
 </tr>
 <tr>
 	<td>assertText</td>
-	<td>xPath=(//div[@class='portlet-msg-info'])[1]</td>
-	<td>There are no tags.</td>
+	<td>//h1[@class='portlet-title']/span[2]</td>
+	<td>Tags Navigation</td>
+</tr>
+<tr>
+	<td>assertText</td>
+	<td>//span[@title='Options']/ul/li/strong/a</td>
+	<td>Options</td>
 </tr>
 <tr>
 	<td>assertVisible</td>
-	<td>//div[2]/section</td>
-	<td>Portlet</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>xPath=(//div[@class='portlet-msg-info'])[2]</td>
-	<td>There are no tags.</td>
+	<td>//img[@title='Minimize']</td>
+	<td>Minimize</td>
 </tr>
 <tr>
 	<td>assertVisible</td>
-	<td>//div[3]/section</td>
-	<td>Portlet</td>
+	<td>//img[@title='Minimize']</td>
+	<td>Minimize</td>
+</tr>
+<tr>
+	<td>assertVisible</td>
+	<td>//img[@title='Remove']</td>
+	<td>Remove</td>
 </tr>
 <tr>
 	<td>assertText</td>
-	<td>xPath=(//div[@class='portlet-msg-info'])[3]</td>
+	<td>//div[@class='portlet-msg-info']</td>
 	<td>There are no tags.</td>
 </tr>
 </tbody></table>
