+++ portal-web/test/functional/com/liferay/portalweb/portlet/tagsnavigation/portlet/addportlettn/ViewPortletTNTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portlet.tagsnavigation.portlet.addportletstn;
+package com.liferay.portalweb.portlet.tagsnavigation.portlet.addportlettn;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,22 +20,23 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class ViewPortletsTNTest extends BaseTestCase {
-	public void testViewPortletsTN() throws Exception {
+public class ViewPortletTNTest extends BaseTestCase {
+	public void testViewPortletTN() throws Exception {
 		selenium.selectWindow("null");
 		selenium.selectFrame("relative=top");
 		selenium.open("/web/guest/home/");
 		selenium.clickAt("link=Tags Navigation Test Page",
 			RuntimeVariables.replace("Tags Navigation Test Page"));
 		selenium.waitForPageToLoad("30000");
-		assertTrue(selenium.isVisible("//div[1]/section"));
+		assertTrue(selenium.isVisible("//section"));
+		assertEquals(RuntimeVariables.replace("Tags Navigation"),
+			selenium.getText("//h1[@class='portlet-title']/span[2]"));
+		assertEquals(RuntimeVariables.replace("Options"),
+			selenium.getText("//span[@title='Options']/ul/li/strong/a"));
+		assertTrue(selenium.isVisible("//img[@title='Minimize']"));
+		assertTrue(selenium.isVisible("//img[@title='Minimize']"));
+		assertTrue(selenium.isVisible("//img[@title='Remove']"));
 		assertEquals(RuntimeVariables.replace("There are no tags."),
-			selenium.getText("xPath=(//div[@class='portlet-msg-info'])[1]"));
-		assertTrue(selenium.isVisible("//div[2]/section"));
-		assertEquals(RuntimeVariables.replace("There are no tags."),
-			selenium.getText("xPath=(//div[@class='portlet-msg-info'])[2]"));
-		assertTrue(selenium.isVisible("//div[3]/section"));
-		assertEquals(RuntimeVariables.replace("There are no tags."),
-			selenium.getText("xPath=(//div[@class='portlet-msg-info'])[3]"));
+			selenium.getText("//div[@class='portlet-msg-info']"));
 	}
 }
